Fixed a typo that accidently made it into grrmud.cc. Updated gen_cmds.spec to include return values for the new weather commands. This is necessary in order to keep the compiler quiet about functions that claim to return values, but don't. Explict downcasting from double->int and float->int in order to make it explicit that this is intentional, and silence compiler warnings.

+++ mud/grrmud/server/object.cc
@@ -493,7 +493,7 @@ int object::getCurWeight() const {
       while ((tmp_obj = cll.next())) {
          tmp_wt += tmp_obj->getCurWeight();
       }//while
-      return (tmp_wt * (this->bag->percentage_weight/100.00)) + extras[5];  // weight of inv, plus container's wt
+      return int((tmp_wt * (this->bag->percentage_weight/100.00)) + extras[5]);  // weight of inv, plus container's wt
    }//else
 }//cur_weight
